Added TourServiceException and TourBookingServiceException. Updated tests.

+++ src/main/java/com/bookinggo/RESTfulDemo/service/TourBookingServiceImpl.java
@@ -3,6 +3,8 @@ package com.bookinggo.RESTfulDemo.service;
 import com.bookinggo.RESTfulDemo.entity.Customer;
 import com.bookinggo.RESTfulDemo.entity.Tour;
 import com.bookinggo.RESTfulDemo.entity.TourBooking;
+import com.bookinggo.RESTfulDemo.exception.TourBookingServiceException;
+import com.bookinggo.RESTfulDemo.exception.TourServiceException;
 import com.bookinggo.RESTfulDemo.repository.CustomerRepository;
 import com.bookinggo.RESTfulDemo.repository.TourBookingRepository;
 import lombok.AllArgsConstructor;
@@ -10,7 +12,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
-import java.sql.SQLException;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
@@ -28,41 +29,50 @@ public class TourBookingServiceImpl implements TourBookingService {
     private final CustomerRepository customerRepository;
 
     @Override
-    public Optional<TourBooking> createBooking(int tourId, Integer customerId, LocalDateTime pickupDateTime, String pickupLocation, Integer participants) throws SQLException {
+    public Optional<TourBooking> createBooking(int tourId, Integer customerId, LocalDateTime pickupDateTime, String pickupLocation, Integer participants) throws TourBookingServiceException {
         log.info("createBooking - tourId: {}, customerId: {}, date: {}, location {}, participants {}", tourId, customerId, pickupDateTime, pickupLocation, participants);
-        final Optional<Tour> tour = tourService.getTourById(tourId);
+        Optional<Tour> tour;
 
-        if (tour.isPresent()) {
-            final Optional<Customer> customer = customerRepository.findById(customerId);
-
-            if (customer.isPresent()) {
-                final TourBooking tourBooking = TourBooking.builder()
-                        .tour(tour.get())
-                        .customer(customer.get())
-                        .pickupDateTime(pickupDateTime)
-                        .pickupLocation(pickupLocation)
-                        .participants(participants)
-                        .build();
-
-                return Optional.of(tourBookingRepository.save(tourBooking));
-            }
-
-            return Optional.empty();
+        try {
+            tour = tourService.getTourById(tourId);
+        } catch (TourServiceException e) {
+            throw new TourBookingServiceException(e.getMessage(), null);
         }
 
-        throw new SQLException("createBooking - failed - tour doesn't exist.");
+        final Optional<Customer> customer = customerRepository.findById(customerId);
+
+        if (customer.isPresent()) {
+            final TourBooking tourBooking = TourBooking.builder()
+                    .tour(tour.get())
+                    .customer(customer.get())
+                    .pickupDateTime(pickupDateTime)
+                    .pickupLocation(pickupLocation)
+                    .participants(participants)
+                    .build();
+
+            return Optional.of(tourBookingRepository.save(tourBooking));
+        }
+
+        throw new TourBookingServiceException("Can't create a booking for this Customer Id. Customer Id doesn't exist.", null);
     }
 
     @Override
-    public Optional<TourBooking> recover(SQLException e, int tourId, Integer customerId, LocalDateTime pickupDateTime, String pickupLocation, Integer participants) {
+    public Optional<TourBooking> recover(TourBookingServiceException e, int tourId, Integer customerId, LocalDateTime pickupDateTime, String pickupLocation, Integer participants) {
         log.info("Called createBooking - recover called instead: {}", e.getMessage());
-        return Optional.empty();
+        throw new TourBookingServiceException(e.getMessage(), null);
     }
 
     @Override
     public List<TourBooking> getBookingsByTourId(int tourId) {
         log.info("getBookingsByTourId - tourId: {}", tourId);
-        return tourBookingRepository.findByTourId(tourId);
+
+        List<TourBooking> tourBookings = tourBookingRepository.findByTourId(tourId);
+
+        if (tourBookings.size() > 0) {
+            return tourBookings;
+        }
+
+        throw new TourBookingServiceException("No Tour Bookings for given Tour Id.", null);
     }
 
     @Override
@@ -93,9 +103,11 @@ public class TourBookingServiceImpl implements TourBookingService {
 
                 return Optional.of(tourBookingRepository.saveAndFlush(bookings.get(0)));
             }
+
+            throw new TourBookingServiceException("Can't update booking. More than one bookings with this Customer Id.", null);
         }
 
-        return Optional.empty();
+        throw new TourBookingServiceException("No Tour Bookings to update for given Customer Id.", null);
     }
 
     @Override
@@ -113,17 +125,16 @@ public class TourBookingServiceImpl implements TourBookingService {
         log.info("deleteAllBookingsWithTourIdAndCustomerId - tourId: {}, customerId: {}", tourId, customerId);
         final List<TourBooking> bookings = tourBookingRepository.findByTourIdAndCustomerId(tourId, customerId);
 
+        if (bookings.size() == 0) {
+            throw new TourBookingServiceException("No Tour Bookings to delete for given Tour Id and Customer Id.", null);
+        }
+
         bookings.forEach(booking -> {
             tourBookingRepository.delete(booking);
             tourBookingRepository.flush();
         });
 
-        final Optional<Customer> customer = customerRepository.findById(customerId);
-        if (customer.isPresent()) {
-            return Optional.of(bookings);
-        }
-
-        return Optional.empty();
+        return Optional.of(bookings);
     }
 
     @Override
@@ -132,17 +143,16 @@ public class TourBookingServiceImpl implements TourBookingService {
         log.info("deleteAllBookingsWithCustomerId - tourId: {}", customerId);
         final List<TourBooking> bookings = tourBookingRepository.findByCustomerId(customerId);
 
+        if (bookings.size() == 0) {
+            throw new TourBookingServiceException("No Tour Bookings to delete for given Customer Id.", null);
+        }
+
         for (TourBooking booking : bookings) {
             tourBookingRepository.delete(booking);
             tourBookingRepository.flush();
         }
 
-        final Optional<Customer> customer = customerRepository.findById(customerId);
-        if (customer.isPresent()) {
-            return Optional.of(bookings);
-        }
-
-        return Optional.empty();
+        return Optional.of(bookings);
     }
 
     @Override
@@ -150,9 +160,11 @@ public class TourBookingServiceImpl implements TourBookingService {
         log.info("deleteAllBookings");
         final List<TourBooking> bookings = tourBookingRepository.findAll();
 
-        if (bookings.size() > 0) {
-            tourBookingRepository.deleteAll();
+        if (bookings.size() == 0) {
+            throw new TourBookingServiceException("No Tour Bookings to delete.", null);
         }
+
+        tourBookingRepository.deleteAll();
 
         return bookings;
     }
